Revamp test suite
Mostly this involves upgrade to the latest mysql_isolated_server, which
is much much much better at playing nice with test/autorun; so we don't
have to have the hacks we had before.

Now we se MiniTest::Unit instead of Test::Unit

Rails 4.0 requires an older version of MiniTest. In that case, we have
to alias the Test class.